requirements.txt
dj-database-url==0.3.0
Django==1.7.3
django-cors-headers==1.0.0
djangorestframework==3.0.4
gunicorn==19.3.0
Pillow==2.8.1
whitenoise==1.0.6
psycopg2==2.5.4